third_party/blink/web_tests/virtual/legacy-client-hints-no-fp-delegation/external/wpt/client-hints/viewport-size-window-different-dimensions.https-expected.txt

This is a testharness.js-based test.
FAIL viewport-size-window-different-dimensions assert_equals: expected "503" but got "FAIL"
Harness: the test ran to completion.
